Refactor doc-comment utilities imports
