+++ src/core/src/comments/doc-comment/service/legacy.ts
@@ -2,7 +2,7 @@ import {
     capitalize,
     isNonEmptyTrimmedString,
     toTrimmedString
-} from "../../../utils/index.js";
+} from "../../../utils/string.js";
 import type { DocCommentLines } from "../../comment-utils.js";
 import { parseDocCommentMetadata, isDocCommentTagLine } from "./metadata.js";
