+++ src/core/src/comments/doc-comment/service/synthetic/merge.ts
@@ -2,13 +2,9 @@ import type {
     DocCommentLines,
     MutableDocCommentLines
 } from "../../../comment-utils.js";
-import {
-    coercePositiveIntegerOption,
-    findLastIndex,
-    toMutableArray,
-    toTrimmedString,
-    isNonEmptyString
-} from "../../../../utils/index.js";
+import { coercePositiveIntegerOption } from "../../../../utils/numeric-options.js";
+import { findLastIndex, toMutableArray } from "../../../../utils/array.js";
+import { isNonEmptyString, toTrimmedString } from "../../../../utils/string.js";
 import { parseDocCommentMetadata } from "../metadata.js";
 import {
     dedupeReturnDocLines,
